requirements.txt

streamlit
pillow
requests
qrcode
pyzbar
opencv-python-headless
